pyproject.toml
[project]
name = "vercel-fastapi-starter"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.121.0",
  "uvicorn>=0.34.0",
  "beautifulsoup4>=4.12.0",
  "httpx>=0.27.0",
  "python-dotenv>=1.0.0",
  "openai>=1.0.0",
]
